Move tests to e2e_new folder
